Tempo Versions

Abstract Tempo Version

text text text

GravityToolsLight.AbstractTempoVersion — Type

AbstractTempoVersion

An abstract type that serves as a base for different implementations of Tempo versions. Implementations must define the get_tempo_command function.

GravityToolsLight.get_tempo_command — Method

get_tempo_command(tempo_version::AbstractTempoVersion)

Obtain the command line call string for the specified Tempo version. This function must be implemented for each specific subtype of AbstractTempoVersion.

Tempo

GravityToolsLight.Tempo — Type

Tempo(custom_data_directory::String = get_tempo_directory())

A structure representing the original Tempo version with an optional custom data directory. If no directory is provided, the default is retrieved from the TEMPO environment variable.

Constructors

Tempo(): Uses the default directory path from TEMPO environment variable.

Tempo(custom_dir::String): Uses a custom directory specified by custom_dir.

GravityToolsLight.get_tempo_directory — Function

get_tempo_directory() -> String

Retrieve the default directory path for Tempo from the TEMPO environment variable. If the variable is not set, it returns a predefined default path.

GravityToolsLight.get_tempo_command — Method

get_tempo_command(tempo_version::Tempo) -> String

Returns the command line call string for the original Tempo version.

Tempo2

GravityToolsLight.Tempo2 — Type

Tempo2(custom_data_directory::String = get_tempo2_directory())

A structure representing the Tempo2 version with an optional custom data directory. If no directory is provided, the default is retrieved from the TEMPO2 environment variable.

Constructors

Tempo2(): Uses the default directory path from TEMPO2 environment variable.

Tempo2(custom_dir::String): Uses a custom directory specified by custom_dir.

GravityToolsLight.get_tempo2_directory — Function

get_tempo2_directory() -> String

Retrieve the default directory path for Tempo2 from the TEMPO2 environment variable. If the variable is not set, it returns a predefined default path.

GravityToolsLight.get_tempo_command — Method

get_tempo_command(tempo_version::Tempo2) -> String

Returns the command line call string for the Tempo2 version.